Fix a 16-bit integer overflow that might occur in statements that use both an EXISTS clause and IN operator with a RHS holding in excess of 32K entries.
FossilOrigin-Name: 65a1f1334d92873ed0b9f2d9ae3e9052091aac19

+++ src/expr.c
@@ -1672,7 +1672,7 @@ void sqlite3CodeSubselect(
       if( sqlite3Select(pParse, pSel, &dest) ){
         return;
       }
-      pExpr->iColumn = (i16)dest.iParm;
+      pExpr->iColumn = dest.iParm;
       ExprSetIrreducible(pExpr);
       break;
     }
